Added solution for day 21.
Today's solution got very verbose and ugly. The data structure could
probably be better modelled to simplify the solution a bit. I also
relied on the unspecified property that no two monkeys wait for the same
monkey to shout. It works for both my input and the example, so it's
probably fine.