Added solution for day 25, part A.
I can't do part B yet since I haven't completed all previous puzzles :(.
I got stuck today and tried to let ChatGPT implement it for me. I
prompted away but the solution was incorrect (though it did manage the
conversion from SNAFU number to decimal after a couple of iterations).

In the end I realized that one needed to implement some kind of carry
digit since it's basically base 5, but 3 and 4 are represented as 1= and
1-.